feat: CALL opcode (#322)
## Pull request type

Please check the type of change your PR introduces:

- [ ] Bugfix
- [X] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

Issue Number: #255 

## What is the new behavior?

Implemented

## Other information

Left some TODOs because the code as it is does not allow for a fully
working implementation. We need to fix RETURN, RETURNDATACOPY and
RETURNDATASIZE, and also to enable gas management.
