+++ src/kakarot/model.cairo
@@ -40,5 +40,6 @@ namespace model {
         intrinsic_gas_cost: felt,
         starknet_contract_address: felt,
         evm_contract_address: felt,
+        parent_context: ExecutionContext*,
     }
 }
