+++ tests/solidity_files/output/Counter.bin
@@ -1,0 +1,1 @@
+608060405234801561001057600080fd5b50600080819055506102ba806100276000396000f3fe608060405234801561001057600080fd5b506004361061004c5760003560e01c806306661abd14610051578063371303c01461006f578063b3bcfa8214610079578063d826f88f14610083575b600080fd5b61005961008d565b604051610066919061012f565b60405180910390f35b610077610093565b005b6100816100ae565b005b61008b61010d565b005b60005481565b60016000808282546100a59190610179565b92505081905550565b60008054116100f2576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016100e990610230565b60405180910390fd5b60016000808282546101049190610250565b92505081905550565b60008081905550565b6000819050919050565b61012981610116565b82525050565b60006020820190506101446000830184610120565b92915050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b600061018482610116565b915061018f83610116565b92508282019050808211156101a7576101a661014a565b5b92915050565b600082825260208201905092915050565b7f636f756e742073686f756c64206265207374726963746c79206772656174657260008201527f207468616e203000000000000000000000000000000000000000000000000000602082015250565b600061021a6027836101ad565b9150610225826101be565b604082019050919050565b600060208201905081810360008301526102498161020d565b9050919050565b600061025b82610116565b915061026683610116565b925082820390508181111561027e5761027d61014a565b5b9291505056fea264697066735822122057b9a31f47fa45e40993e34b410988fdf7bda73b70c769e03d1a4fc0538bc7d564736f6c63430008110033
