+++ tests/units/test_account_registry.py
@@ -9,13 +9,13 @@ random.seed(0)
 @pytest.mark.asyncio
 class TestAccountRegistry:
     async def test_should_set_starknet_and_evm_contract_addresses(
-        self, account_registry: StarknetContract
+        self, account_registry: StarknetContract, kakarot: StarknetContract
     ):
         starknet_contract_address = random.randint(0, 10_000)
         evm_contract_address = random.randint(0, 10_000)
         await account_registry.set_account_entry(
             starknet_contract_address, evm_contract_address
-        ).execute(caller_address=1)
+        ).execute(caller_address=kakarot.contract_address)
         stored_starknet_contract_address = (
             await account_registry.get_starknet_contract_address(
                 evm_contract_address
